requirements.txt
bcrypt==3.2.0
cffi==1.14.4
click==7.1.2
cryptography==3.3.1
ecdsa==0.14.1
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
psycopg2==2.8.6
pyasn1==0.4.8
pycparser==2.20
python-jose==3.2.0
rsa==4.7
six==1.15.0
Werkzeug==1.0.1
